app/requirements.txt
certifi==2025.10.5
charset-normalizer==3.4.4
idna==3.11
influxdb-client==1.49.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
reactivex==4.0.4
requests==2.32.5
setuptools==80.9.0
six==1.17.0
typing_extensions==4.15.0
urllib3==2.5.0
